rename main component files
